掲示板

+++ resources/views/show.blade.php
@@ -7,13 +7,15 @@
         <p>{{ $article->user_name }}</p>
         <p>
             <a href={{ route('article.list') }} class='btn btn-outline-primary'>一覧に戻る</a>
-            <a href={{ route('article.edit',['id' =>$article->id]) }} class='btn btn-outline-primary'>編集</a>
         </p>
         <div>
-            {{ Form::open(['method' => 'delete', 'route' => ['article.delete', $article->id]]) }}
-                {{ Form::submit('削除',['class'=>'btn btn-outline-secondary']) }}
-            {{ Form::close() }}
-
-            
+            @auth 
+              @if ($article->user_id ===$login_user_id)
+                    {{ Form::open(['method' => 'delete', 'route' => ['article.delete', $article->id]]) }}
+                        {{ Form::submit('削除',['class'=>'btn btn-outline-secondary']) }}
+                        <a href={{ route('article.edit',['id' =>$article->id]) }} class='btn btn-outline-primary'>編集</a>
+                    {{ Form::close() }}
+              @endif
+            @endauth
         </div>
 @endsection
